src/main/java/com/recarga/pay/wallet/services/impl/UserServiceImpl.java
package com.recarga.pay.wallet.services.impl;

import com.recarga.pay.wallet.entities.User;
import com.recarga.pay.wallet.repositories.UserRepository;
import com.recarga.pay.wallet.services.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class UserServiceImpl implements UserService {

    @Autowired
    private UserRepository userRepository;

    @Override
    public User createUser(String name) {
        User user = new User();
        user.setName(name);
        return userRepository.save(user);
    }

    @Override
    public Optional<User> getUserById(Long id) {
        return userRepository.findById(id);
    }
}
